// frontend/src/types/Car.ts
export interface Car {
  id: number;
  make: string;
  model: string;
  variant: string;
  color: string;
  fuel_type: string;
  registration_number: string;
  registration_date: string;
  rc_status: string;
  rc_expiry_date: string;
  chassis_number: string;
  engine_number: string;
  manufacturing_year: number;
  manufacturing_month: number;
  owner_serial_number: number;
  mileage: number;
  cc: string;
  emission_norms: string;
  transmission: string;
  category: string;
  insurance_type: string;
  insurance_expiry: string;
  price: string;
  ready_for_sales: string;
 
}